V1 of Frontier Code
BUG: The A* algorithm seems to have a bug - it is currently sending the robot through walls.

Map Generation: Commented out line at the bottom. They were running and causing the code to hang.

Frontier2: takes in frontier values, calls A*, chooses best path and returns path. Need to continue tuning this so the algorithm doesn't just choose the next closest frontier (which is basically what it is doing now)

Robot: Updated to include observed map and frontier values. Also now takes in a location and returns the next action to get to that location

Testing: Reconfigured testing to interface with frontier algorithm. Loops through path returned from frontier and updates the robot and world at each time step.

Underground: Observations are now of fidelity (frontier) value at each step, instead of a 1 or 0 for the tunnel. Should work the same because there is either a value there (frontier) or none (wall)